# services/api/docs/swagger.yaml
definitions:
...
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
...
  model.Ayah:
    properties:
      default_language:
        type: string
      hizb_quarter:
        type: integer
      id:
        type: integer
      juz_id:
        type: integer
      juz_number:
        type: integer
      manzil:
        type: integer
      media:
        items:
          $ref: '#/definitions/model.AyahAsset'
        type: array
      number:
        type: integer
      page:
        type: integer
      ruku:
        type: integer
      sajda:
        type: boolean
      surah:
        $ref: '#/definitions/model.Surah'
      surah_id:
        type: integer
      translation:
        $ref: '#/definitions/model.Translation'
      translation_id:
        type: integer
    type: object
  model.AyahAsset:
    properties:
      ayah_id:
        type: integer
      id:
        type: integer
      multimedia:
        $ref: '#/definitions/model.Multimedia'
      multimedia_id:
        type: integer
    type: object
...
  model.Juz:
    properties:
      ayahs:
        items:
          $ref: '#/definitions/model.Ayah'
        type: array
      end_ayah:
        $ref: '#/definitions/model.Ayah'
      end_ayah_id:
        type: integer
      end_surah:
        $ref: '#/definitions/model.Surah'
      end_surah_id:
        type: integer
      id:
        type: integer
      media:
        items:
          $ref: '#/definitions/model.JuzAsset'
        type: array
      number:
        type: integer
      start_ayah:
        $ref: '#/definitions/model.Ayah'
      start_ayah_id:
        type: integer
      start_surah:
        $ref: '#/definitions/model.Surah'
      start_surah_id:
        type: integer
      total_ayah:
        type: integer
    type: object
  model.JuzAsset:
    properties:
      id:
        type: integer
      juz_id:
        type: integer
      multimedia:
        $ref: '#/definitions/model.Multimedia'
      multimedia_id:
        type: integer
    type: object
...
  model.Multimedia:
    properties:
      file_name:
        description: File Name
        type: string
      file_size:
        description: File Size
        format: float
        type: number
      format:
        description: Format
        type: string
      id:
        type: integer
      original_file_name:
        description: Original File Name
        type: string
      title:
        description: Title
        type: string
      translation:
        allOf:
        - $ref: '#/definitions/model.Translation'
        description: Translation
      translation_id:
        description: Translation ID
        type: integer
      url:
        description: Url
        example: protocol://domain.ltd/path/to/image.extension
        type: string
    type: object
...
  model.Surah:
    properties:
      ayahs:
        items:
          $ref: '#/definitions/model.Ayah'
        type: array
      default_language:
        type: string
      end_juz:
        $ref: '#/definitions/model.Juz'
      id:
        type: integer
      identifier:
        type: string
      media:
        items:
          $ref: '#/definitions/model.SurahAsset'
        type: array
      next_surah:
        $ref: '#/definitions/model.Surah'
      number:
        type: integer
      number_of_ayahs:
        type: integer
      prev_surah:
        $ref: '#/definitions/model.Surah'
      revelation_type:
        type: string
      slug:
        type: string
      start_juz:
        $ref: '#/definitions/model.Juz'
      translation:
        $ref: '#/definitions/model.Translation'
      translation_id:
        type: integer
    type: object
  model.SurahAsset:
    properties:
      id:
        type: integer
      multimedia:
        $ref: '#/definitions/model.Multimedia'
      multimedia_id:
        type: integer
      surah_id:
        type: integer
    type: object
...
  model.Translation:
    properties:
      ar:
        type: string
      ar_format:
        type: string
      ar_html:
        type: string
      ar_waqaf:
        type: string
      description_ar:
        type: string
      description_en:
        type: string
      description_idn:
        type: string
      en:
        type: string
      id:
        type: integer
      idn:
        type: string
      latin_en:
        type: string
      latin_idn:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/ayah:
    get:
      consumes:
      - application/json
      description: Get paginated list of all ayah
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
      summary: List all ayah
      tags:
      - Ayah
    post:
      consumes:
      - application/json
      description: Create a new ayah entry
      parameters:
      - description: Ayah data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Ayah'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Create ayah
      tags:
      - Ayah
  /api/v1/ayah/{id}:
    get:
      consumes:
      - application/json
      description: Get a single ayah by its ID
      parameters:
      - description: Ayah ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get ayah by ID
      tags:
      - Ayah
    put:
      consumes:
      - application/json
      description: Update ayah by ID
      parameters:
      - description: Ayah ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ayah data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Ayah'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Update ayah
      tags:
      - Ayah
  /api/v1/ayah/{id}/{scoped}:
    delete:
      consumes:
      - application/json
      description: Soft delete ayah by ID and scope
      parameters:
      - description: Ayah ID
        in: path
        name: id
        required: true
        type: integer
      - description: Delete scope (soft/hard)
        in: path
        name: scoped
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Delete ayah
      tags:
      - Ayah
  /api/v1/ayah/daily:
    get:
      consumes:
      - application/json
      description: Get a random ayah for daily reading
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get daily ayah
      tags:
      - Ayah
  /api/v1/ayah/hizb/{hizb}:
    get:
      consumes:
      - application/json
      description: Get all ayah in a specific hizb quarter (1-240)
      parameters:
      - description: Hizb quarter number
        in: path
        name: hizb
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get ayah by hizb quarter
      tags:
      - Ayah
  /api/v1/ayah/keyset:
    get:
      consumes:
      - application/json
      description: Get paginated list of ayah using keyset pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: List ayah with keyset pagination
      tags:
      - Ayah
  /api/v1/ayah/number/{number}:
    get:
      consumes:
      - application/json
      description: Get ayah by its number within a surah
      parameters:
      - description: Ayah number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get ayah by number
      tags:
      - Ayah
  /api/v1/ayah/page/{page}:
    get:
      consumes:
      - application/json
      description: Get all ayah on a specific Quran page (1-604)
      parameters:
      - description: Page number
        in: path
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get ayah by page
      tags:
      - Ayah
  /api/v1/ayah/surah/number/{number}:
    get:
      consumes:
      - application/json
      description: Get all ayah for a specific surah number
      parameters:
      - description: Surah number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get ayah by surah number
      tags:
      - Ayah
...
swagger: "2.0"